Altered the math helper methods to take care of some warnings.

+++ NinjaAttackTutorial/UI Elemnts/SharredScene.swift
@@ -124,7 +124,7 @@ class SharredScene: SKScene {
     //===================
     
     func random() -> CGFloat {
-        return CGFloat(Float(arc4random()) / 0xFFFFFFFF)
+        return CGFloat(Double(arc4random()) / 0xFFFFFFFF)
     }
     
     func random(min: CGFloat, max: CGFloat) -> CGFloat {
